Split SQLite encoding implementation
